Abstracts ruletable to component, sets rules/actions lists to consume
Fixes https://projects.engineering.redhat.com/browse/RHIADVISOR-450
