+++ src/SmartComponents/Actions/_actions.scss
@@ -1,24 +1,5 @@
-@import '~@red-hat-insights/insights-frontend-components/Utilities/_variables';
-@import '~@red-hat-insights/insights-frontend-components/Utilities/_mixins';
-
 .page__actions {
-    .actions__view--title { text-transform: capitalize; }
-    .actions__view {
-        table.rules-table {
-            th:first-child { width: 50%; }
-            td {
-                padding: $ins-padding;
-                &:not(:first-of-type) { text-align: center; }
-            }
-            tfoot tr td:first-child { padding: 0; }
-        }
-    }
-}
-
-.actions__description { margin: $ins-margin 0; }
-
-@media only screen and (min-width: 768px) {
-    .ins-l-icon-group__vertical > *:not(:first-child) {
-        @include rem('margin-top', 5px);
-    }
+  .actions__view--title {
+    text-transform: capitalize;
+  }
 }
